# api-spec.yaml
openapi: 3.0.3
info:
  title: Buchmanagement API
  version: "1.0.0"
  description: |
    Diese API ermöglicht die Verwaltung von Benutzern, deren Büchern und Buchrezensionen.
    Bücher werden anhand ihrer ISBN hinzugefügt, und die Buchdetails werden automatisch
    über die OpenLibrary API abgerufen.
servers:
  - url: http://localhost:8080
    description: Basis-URL ohne Pfad-Prefix
paths:
...
  /users/{userId}/books/{isbn}/reviews:
    get:
      summary: Alle Rezensionen eines Buches
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
        - in: path
          name: isbn
          schema:
            type: string
          required: true
      responses:
        "201":
          description: Rezension erstellt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Ungültige Eingabedaten (z. B. Rating außerhalb des erlaubten Bereichs).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Benutzer oder Buch nicht gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    post:
      summary: Rezension hinzufügen
      description: Erstellt Bewertung und Text für ein Buch.
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
        - in: path
          name: isbn
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: integer
                reviewText:
                  type: string
              required:
                - rating
                - reviewText
      responses:
        "201":
          description: Rezension erstellt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Ungültige Eingabedaten (z. B. Rating außerhalb des erlaubten Bereichs).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Benutzer oder Buch nicht gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /users/{userId}/books/{isbn}/reviews/{reviewId}:
    put:
      summary: Rezension aktualisieren
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
        - in: path
          name: isbn
          schema:
            type: string
          required: true
        - in: path
          name: reviewId
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: integer
                reviewText:
                  type: string
      responses:
        "201":
          description: Rezension erstellt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Ungültige Eingabedaten (z. B. Rating außerhalb des erlaubten Bereichs).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Benutzer oder Buch nicht gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    delete:
      summary: Rezension löschen
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
        - in: path
          name: isbn
          schema:
            type: string
          required: true
        - in: path
          name: reviewId
          schema:
            type: string
          required: true
      responses:
        "204":
          description: Rezension gelöscht.
        "404":
          description: Benutzer, Buch oder Rezension nicht gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
components:
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
...
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rating:
          type: integer
        reviewText:
          type: string
        bookId:
          type: string
          format: uuid
      required:
        - id
        - rating
        - reviewText